第三次提交

+++ userSystem/src/main/java/com/dog/usersystem/service/Impl/UserServiceImpl.java
@@ -1,19 +1,20 @@
 package com.dog.usersystem.service.Impl;
 
+import com.Dog.Doman.Result;
+import com.Dog.Doman.ResultEnum;
+import com.Dog.Exception.BusinessException;
+import com.Dog.Utils.JwtUtil;
+import com.Dog.Utils.RedisUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.dog.usersystem.dao.UserMapper;
-import com.dog.usersystem.doman.Result;
-import com.dog.usersystem.doman.po.User;
+import com.dog.usersystem.doman.dto.User;
 import com.dog.usersystem.doman.vo.req.UserLoginReq;
 import com.dog.usersystem.doman.vo.req.UserRegisterReq;
 import com.dog.usersystem.doman.vo.resp.JwtTokenResp;
 import com.dog.usersystem.service.UserService;
-import com.dog.usersystem.utils.JwtUtil;
-import com.dog.usersystem.utils.RedisUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -32,49 +33,38 @@ public class UserServiceImpl implements UserService {
     private static final int USERNAME_EXPIRE = 900;
 
     @Override
-    public ResponseEntity<Result<Void>> registerUser(UserRegisterReq userRegisterReq) {
-        // 记录日志
-        log.info("【用户注册】{}", userRegisterReq);
-
+    public Result<Void> registerUser(UserRegisterReq userRegisterReq) {
         // 检查用户名是否存在
         if (userMapper.selectOne(new QueryWrapper<User>().eq("username", userRegisterReq.getUsername())) != null) {
-            log.warn("【用户名已存在】");
-            return ResponseEntity.status(500).body(Result.error());
+            throw new BusinessException(ResultEnum.USERNAME_EXISTS);
         }
 
         // 密码加密
         String newPassword = passwordEncoder.encode(userRegisterReq.getPassword());
-        userRegisterReq.setPassword(newPassword);
 
         // 存入数据库
         User user = new User();
         BeanUtils.copyProperties(userRegisterReq, user);
+        user.setPassword(newPassword);
         userMapper.insert(user);
 
-        // 记录日志, 响应结果
-        log.info("【注册成功】");
-        return ResponseEntity.ok(Result.success());
+        return Result.success();
     }
 
     @Override
-    public ResponseEntity<Result<JwtTokenResp>> loginUser(UserLoginReq userLoginReq) {
-        // 记录日志
-        log.info("【用户登录】{}", userLoginReq);
-
+    public Result<JwtTokenResp> loginUser(UserLoginReq userLoginReq) {
         // 查询用户名
         User user = userMapper.selectOne(new QueryWrapper<User>().eq("username", userLoginReq.getUsername()));
 
         // 检验用户有效性
         if (user == null) {
-            log.warn("【用户名不存在】");
-            return ResponseEntity.status(500).body(Result.error());
+            throw new BusinessException(ResultEnum.USERNAME_NOT_EXITS);
         }
 
         // 检验密码(非明文)
         if (!passwordEncoder.matches(userLoginReq.getPassword(), user.getPassword())) {
-            log.warn("【用户密码错误】");
-            redisUtil.setUserBlackList(userLoginReq.getUsername(), USERNAME_EXPIRE);
-            return ResponseEntity.status(500).body(Result.error());
+            redisUtil.setUsernameBlackList(userLoginReq.getUsername(), USERNAME_EXPIRE);
+            throw new BusinessException(ResultEnum.PASSWORD_INCORRECT);
         }
 
         // 生成 JWT 令牌
@@ -86,36 +76,33 @@ public class UserServiceImpl implements UserService {
         redisUtil.set("user_access_" + user.getUserId(), jwtTokenResp.getAccessToken(), JwtUtil.ACCESS_EXPIRE / 1000);
         redisUtil.set("user_refresh_" + user.getUserId(),  jwtTokenResp.getRefreshToken(), JwtUtil.REFRESH_EXPIRE / 1000);
 
-        // 记录日志, 响应结果
-        log.info("【登陆成功】");
         redisUtil.deleteUsernameBlackList(userLoginReq.getUsername());
-        return ResponseEntity.ok(Result.success(jwtTokenResp));
+        return Result.success(jwtTokenResp);
     }
 
     @Override
-    public ResponseEntity<Result<Void>> logoutUser(Long userId) {
-        // 记录日志
-        log.info("【用户登出】");
-
+    public Result<Void> logoutUser(Long userId) {
         // 查token
-        String accessToken = redisUtil.get("user_access_" + userId).toString();
-        String refreshToken = redisUtil.get("user_refresh_" + userId).toString();
+        String accessToken = (String) redisUtil.get("user_access_" + userId);
+        String refreshToken = (String) redisUtil.get("user_refresh_" + userId);
 
         // token 加入黑名单
-        redisUtil.setBlackList(accessToken, JwtUtil.getRemainingExpireSeconds(accessToken));
-        redisUtil.setBlackList(refreshToken, JwtUtil.getRemainingExpireSeconds(refreshToken));
+        if (accessToken != null) {
+            redisUtil.setBlackList(accessToken, JwtUtil.getRemainingExpireSeconds(accessToken));
+            redisUtil.delete("user_access_" + userId);
+        }
+        if (refreshToken != null) {
+            redisUtil.setBlackList(refreshToken, JwtUtil.getRemainingExpireSeconds(refreshToken));
+            redisUtil.delete("user_refresh_" + userId);
+        }
 
-        // 记录日志, 响应结果
-        log.info("【登出成功】");
-        redisUtil.delete("user_refresh_" + userId);
-        return ResponseEntity.ok(Result.success());
+        return Result.success();
     }
 
     @Override
-    public ResponseEntity<Result<JwtTokenResp>> refreshToken(Long userId, String refreshToken) {
-        // 记录日志
-        log.info("【刷新 Token】");
+    public Result<JwtTokenResp> refreshToken(Long userId) {
         String accessToken = JwtUtil.createAccessToken(userId);
+        String refreshToken = JwtUtil.createRefreshToken(userId);
 
         // 刷新 redis
         redisUtil.set("user_access_" + userId, accessToken, JwtUtil.ACCESS_EXPIRE / 1000);
@@ -125,8 +112,6 @@ public class UserServiceImpl implements UserService {
         jwtTokenResp.setAccessToken(accessToken);
         jwtTokenResp.setRefreshToken(refreshToken);
 
-        // 记录日志, 响应结果
-        log.info("【刷新成功】");
-        return ResponseEntity.ok(Result.success(jwtTokenResp));
+        return Result.success(jwtTokenResp);
     }
 }
